refactor(allocator): extract admin HTML pages into routes/admin_pages.py
Moves the eight thin admin-page handlers (create_instances, admin,
byo_onboarding, view_instances, view_instances_fragment, delete_instances,
get_vm_logs, scheduled_destruction_page) out of main.py into a new
admin_pages blueprint, following the pattern established by
routes/public.py. main.py now registers six blueprints alphabetically.
